+++ lib/anycable/rails/actioncable/connection/persistent_session.rb
@@ -28,3 +28,7 @@ module ActionCable
     end
   end
 end
+
+::ActionCable::Connection::Base.prepend(
+  ::ActionCable::Connection::PersistentSession
+)